+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.3.4 - Upcoming Release
 
+- Ruler now displays time units
 - Added Feature to change time units
 - Scrolling Mode is now a global user setting
 - Viewer reloads previous state upon closing and reopening VScode